+++ src/Aks/Aks/ChangeLog.md
@@ -28,6 +28,7 @@
 * Added parameter `-EnableFIPS` for `New-AzAksCluster` and `New-AzAksNodePool`
 * Added parameter `-AutoScalerProfile` for `New-AzAksCluster` and `Set-AzAksCluster`
 * Added parameter `-GpuInstanceProfile` for `New-AzAksCluster` and `New-AzAksNodePool`
+* Added parameter `-EnableUptimeSLA` for `New-AzAksCluster` and `Set-AzAksCluster`
 
 ## Version 5.1.0
 * Bumped API version to 2022-09-01